mobile/src/scanner/rfid.rs
use std::sync::Arc;
use tokio::sync::mpsc;
use serde::{Serialize, Deserialize};
use crate::mesh::error::MeshError;
use crate::offline::storage::{MobileStorage, MobileData, MobileSyncStatus};
use crate::sync::SyncPriority;

// RFID specific imports
use serialport::{SerialPort, SerialPortSettings};
use bytes::{BytesMut, BufMut};

#[derive(Debug, Clone, Serialize, Deserialize)]
pub struct RFIDData {
    pub tag_id: String,
    pub timestamp: chrono::DateTime<chrono::Utc>,
    pub tag_type: RFIDTagType,
    pub data: Vec<u8>,
    pub security_level: SecurityLevel,
}

#[derive(Debug, Clone, Serialize, Deserialize)]
pub enum RFIDTagType {
    LowFrequency,    // 125-134.2 kHz
    HighFrequency,   // 13.56 MHz
    UltraHighFrequency, // 856-960 MHz
}

#[derive(Debug, Clone, Serialize, Deserialize)]
pub enum SecurityLevel {
    None,
    Password,
    Encrypted,
    SecureElement,
}

pub struct RFIDScanner {
    storage: Arc<MobileStorage>,
    scan_channel: mpsc::Sender<RFIDData>,
    reader_config: RFIDReaderConfig,
    port: Option<Box<dyn SerialPort>>,
}

#[derive(Debug, Clone)]
pub struct RFIDReaderConfig {
    pub power_level: u8,
    pub frequency: u32,
    pub timeout: std::time::Duration,
    pub port_name: String,
    pub baud_rate: u32,
}

impl RFIDScanner {
    pub fn new(storage: Arc<MobileStorage>, config: RFIDReaderConfig) -> Self {
        let (tx, _) = mpsc::channel(100);
        Self {
            storage,
            scan_channel: tx,
            reader_config: config,
            port: None,
        }
    }

    pub async fn start_scanning(&self) -> Result<(), MeshError> {
        self.initialize_reader().await?;
        self.process_scans().await?;
        Ok(())
    }

    async fn initialize_reader(&mut self) -> Result<(), MeshError> {
        let settings = SerialPortSettings {
            baud_rate: self.reader_config.baud_rate,
            data_bits: serialport::DataBits::Eight,
            flow_control: serialport::FlowControl::None,
            parity: serialport::Parity::None,
            stop_bits: serialport::StopBits::One,
            timeout: self.reader_config.timeout,
        };

        let port = serialport::open_with_settings(&self.reader_config.port_name, &settings)
            .map_err(|e| MeshError::SystemError(format!("Failed to open RFID reader port: {}", e)))?;

        // Configure reader
        self.configure_reader(&port).await?;
        self.port = Some(port);

        Ok(())
    }

    async fn configure_reader(&self, port: &Box<dyn SerialPort>) -> Result<(), MeshError> {
        // Send configuration commands to the RFID reader
        let mut command = BytesMut::with_capacity(8);
        
        // Set power level
        command.put_u8(0x02); // Start of command
        command.put_u8(0x01); // Power level command
        command.put_u8(self.reader_config.power_level);
        command.put_u8(0x03); // End of command
        
        port.write_all(&command)
            .map_err(|e| MeshError::SystemError(format!("Failed to configure reader: {}", e)))?;

        // Set frequency
        command.clear();
        command.put_u8(0x02);
        command.put_u8(0x02); // Frequency command
        command.put_u32(self.reader_config.frequency);
        command.put_u8(0x03);
        
        port.write_all(&command)
            .map_err(|e| MeshError::SystemError(format!("Failed to configure reader: {}", e)))?;

        Ok(())
    }

    async fn process_scans(&self) -> Result<(), MeshError> {
        let storage = self.storage.clone();
        let scan_channel = self.scan_channel.clone();
        let config = self.reader_config.clone();
        let port = self.port.as_ref()
            .ok_or_else(|| MeshError::SystemError("RFID reader not initialized".to_string()))?
            .try_clone()
            .map_err(|e| MeshError::SystemError(format!("Failed to clone port: {}", e)))?;

        tokio::spawn(async move {
            loop {
                match Self::scan_next(&port, &config).await {
                    Ok(rfid_data) => {
                        if let Ok(verified_data) = Self::verify_rfid_data(&rfid_data).await {
                            let mobile_data = MobileData {
                                id: verified_data.tag_id.clone(),
                                data_type: "rfid_tag".to_string(),
                                content: serde_json::to_value(verified_data.clone()).unwrap(),
                                timestamp: chrono::Utc::now(),
                                sync_status: MobileSyncStatus::Pending,
                                priority: SyncPriority::High,
                            };

                            if let Err(e) = storage.store(mobile_data).await {
                                eprintln!("Failed to store RFID data: {}", e);
                            }

                            let _ = scan_channel.send(verified_data).await;
                        }
                    }
                    Err(e) => eprintln!("RFID scan error: {}", e),
                }
                tokio::time::sleep(std::time::Duration::from_millis(100)).await;
            }
        });

        Ok(())
    }

    async fn scan_next(port: &Box<dyn SerialPort>, config: &RFIDReaderConfig) -> Result<RFIDData, MeshError> {
        let mut buffer = [0u8; 256];
        let mut tag_data = BytesMut::new();

        // Send inventory command
        let inventory_cmd = [0x02, 0x03, 0x01, 0x03]; // Example command format
        port.write_all(&inventory_cmd)
            .map_err(|e| MeshError::SystemError(format!("Failed to send inventory command: {}", e)))?;

        // Read response
        let bytes_read = port.read(&mut buffer)
            .map_err(|e| MeshError::SystemError(format!("Failed to read RFID data: {}", e)))?;

        if bytes_read > 0 {
            tag_data.extend_from_slice(&buffer[..bytes_read]);
            
            // Parse RFID data
            let tag_id = hex::encode(&tag_data[1..13]); // Example parsing
            let tag_type = Self::determine_tag_type(&tag_data)?;
            
            return Ok(RFIDData {
                tag_id,
                timestamp: chrono::Utc::now(),
                tag_type,
                data: tag_data.to_vec(),
                security_level: SecurityLevel::None, // Determine based on tag data
            });
        }

        Err(MeshError::SystemError("No RFID tag detected".to_string()))
    }

    fn determine_tag_type(data: &[u8]) -> Result<RFIDTagType, MeshError> {
        // Determine tag type based on protocol identifier or frequency
        match data[0] {
            0x01 => Ok(RFIDTagType::LowFrequency),
            0x02 => Ok(RFIDTagType::HighFrequency),
            0x03 => Ok(RFIDTagType::UltraHighFrequency),
            _ => Err(MeshError::SystemError("Unknown RFID tag type".to_string())),
        }
    }

    async fn verify_rfid_data(data: &RFIDData) -> Result<RFIDData, MeshError> {
        match data.security_level {
            SecurityLevel::None => Ok(data.clone()),
            SecurityLevel::Password => {
                // Verify password protection
                Self::verify_password(data).await?;
                Ok(data.clone())
            }
            SecurityLevel::Encrypted => {
                // Decrypt and verify data
                Self::decrypt_data(data).await?;
                Ok(data.clone())
            }
            SecurityLevel::SecureElement => {
                // Verify with secure element
                Self::verify_secure_element(data).await?;
                Ok(data.clone())
            }
        }
    }

    async fn verify_password(data: &RFIDData) -> Result<(), MeshError> {
        // Implement password verification
        Ok(())
    }

    async fn decrypt_data(data: &RFIDData) -> Result<(), MeshError> {
        // Implement data decryption
        Ok(())
    }

    async fn verify_secure_element(data: &RFIDData) -> Result<(), MeshError> {
        // Implement secure element verification
        Ok(())
    }
} 